turning into a functional.
