--- examples/pitch_dectection/pitch.py
@@ -1,143 +1,0 @@
-import math
-
-import torch
-import torchaudio
-
-
-def compute_nccf(waveform, sample_rate, frame_time=10 ** -2, freq_low=85):
-    """
-    Compute Normalized Cross-Correlation Function (NCCF).
-    """
-    EPSILON = 10 ** (-9)
-
-    # Number of lags to check
-    # Empirically determined default lower voice frequency threshold
-    # based on fundamental frequency for human voice: 85-180 Hz or 165-255 Hz
-    # https://en.wikipedia.org/wiki/Voice_frequency
-    lags = math.ceil(sample_rate / freq_low)
-
-    frame_size = int(math.ceil(sample_rate * frame_time))
-
-    waveform_length = waveform.size()[-1]
-    num_of_frames = math.ceil(waveform_length / frame_size)
-
-    p = lags + num_of_frames * frame_size - waveform_length
-    waveform = torch.nn.functional.pad(waveform, (0, p))
-
-    # Compute lags
-    output_lag = []
-    for lag in range(1, lags + 1):
-        s1 = waveform[..., :-lag].unfold(-1, frame_size, frame_size)[
-            ..., :num_of_frames, :
-        ]
-        s2 = waveform[..., lag:].unfold(-1, frame_size, frame_size)[
-            ..., :num_of_frames, :
-        ]
-
-        output_frames = (
-            (s1 * s2).sum(-1)
-            / (EPSILON + s1.norm(dim=-1))
-            / (EPSILON + s2.norm(dim=-1))
-        )
-
-        output_lag.append(output_frames.unsqueeze(-1))
-
-    nccf = torch.cat(output_lag, -1)
-
-    return nccf
-
-
-def _combine_max(a, b, thresh=0.99):
-    """
-    Take value from first if bigger than a multiplicative factor of the second, elementwise.
-    """
-    mask = (a[0] > thresh * b[0]).to(int)
-    values = mask * a[0] + (1 - mask) * b[0]
-    indices = mask * a[1] + (1 - mask) * b[1]
-    return values, indices
-
-
-def find_max_per_frame(nccf, sample_rate, freq_high=3400):
-    """
-    For each frame, take the highest value of NCCF,
-    apply centered median smoothing, and convert to frequency.
-
-    Note: If the max among all the lags is very close
-    to the first half of lags, then the latter is taken.
-    """
-
-    # Empirically determined default upper voice frequency threshold
-    # based on usable voice frequencies for telephony: 30-3400 Hz
-    # https://en.wikipedia.org/wiki/Voice_frequency
-    lag_min = math.ceil(sample_rate / freq_high)
-
-    # Find near enough max that is smallest
-
-    best = torch.max(nccf[..., lag_min:], -1)
-
-    half_size = nccf.shape[-1] // 2
-    half = torch.max(nccf[..., lag_min:half_size], -1)
-
-    best = _combine_max(half, best)
-    indices = best[1]
-
-    # Add back minimal lag
-    indices += lag_min
-    # Add 1 empirical calibration offset
-    indices += 1
-
-    return indices
-
-
-def median_smoothing(indices, window_length=30):
-    """
-    Apply median smoothing to the 1D tensor over the given window.
-    """
-
-    # Centered windowed
-    pad_length = (window_length - 1) // 2
-
-    # "replicate" padding in any dimension
-    indices = torch.nn.functional.pad(
-        indices, (pad_length, 0), mode="constant", value=0
-    )
-
-    idx = [1] * indices.dim()
-    idx[-1] = pad_length
-    indices[..., :pad_length] = indices[..., pad_length].unsqueeze(-1).repeat(*idx)
-    roll = indices.unfold(-1, window_length, 1)
-
-    values, _ = torch.median(roll, -1)
-    return values
-
-
-def convert_indices_to_freq(indices, sample_rate):
-    EPSILON = 10 ** (-9)
-    freq = sample_rate / (EPSILON + indices.to(torch.float))
-    return freq
-
-
-if __name__ == "__main__":
-
-    # Files from https://www.mediacollege.com/audio/tone/download/
-    tests = [
-        ("100Hz_44100Hz_16bit_05sec.wav", 100),
-        ("440Hz_44100Hz_16bit_05sec.wav", 440),
-    ]
-
-    for filename, freq_ref in tests:
-        waveform, sample_rate = torchaudio.load(filename)
-
-        # Convert to stereo for testing purposes
-        waveform = waveform.repeat(2, 1, 1)
-
-        nccf = compute_nccf(waveform, sample_rate)
-        indices = find_max_per_frame(nccf, sample_rate)
-        indices = median_smoothing(indices)
-        freq = convert_indices_to_freq(indices, sample_rate)
-
-        threshold = 1
-        if ((freq - freq_ref).abs() > threshold).sum():
-            print("Test failed with " + filename)
-        else:
-            print("Test passed with " + filename)
